fix addSinesPow definition

+++ Gamma/tbl.h
@@ -134,7 +134,7 @@ void inline addSines(ArrayType<T,Alloc>& dst, const A * amps, const C * cycs, in
 /// @param[in] wphs		phase of composite waveform, in [0,1]
 template <int InvPower, class T>
 void addSinesPow(
-	T * dst, uint32_t len, double numh,
+	T * dst, uint32_t len, int numh,
 	double hmul=1, double hshf=1, double amp=1, double hphs=0, double wphs=0
 );
 
@@ -560,8 +560,8 @@ void addSine(T * dst, uint32_t len, double cycles, double amp, double phs){
 
 template <int InvPower, class T>
 void addSinesPow(
-	T * dst, uint32_t len,
-	double numh, double hmul, double hshf, double amp, double hphs, double wphs
+	T * dst, uint32_t len, int numh,
+	double hmul, double hshf, double amp, double hphs, double wphs
 ){
 	const double inc1 = M_2PI / len;
